Add MechTextBuffer
This object stores text and offers iterators for iteration itself
and content manipulation. Internally the text is stored as a list
of refcounted elements, and a GSequence of pointers to locations
in these strings wrapping it. When all references in the GSequence
drop on one of these stored strings, the chunk is then freed.