textbuffer: Make it able to store user data
All text in a GSequence section is guaranteed to have all the
same data. When adding or removing data, the sections are split
or merged conveniently if possible.

The data, when registered, requires a GObject instance pointer,
the returned ID and all data set through it is then tied to
that instance, this will be used to implement per-view specific
data.