Add MechTextView
This area implements MechText. Through the MechView methods, it
can keep track of the portions that are currently visible, only
keeping pango resources on these.

The total/viewport height is found out through heuristics and updated
often as scrolling happens, this may induce changes on the scrollbar
handle size, but requires no background quests for the total size.