surface: Store a history of damaged areas
The history list length is hardcoded to 3 to allow for up to
triple-buffered surfaces, the damaged areas are stored in device
space coordinates and transformed back to user space coordinates
when composing the clip mask.

At the time of rendering, the clip region is composed as an union
of all damaged areas since the surface was last used. If the buffer
age is too new (0) or too old (> 3), the whole surface is damaged
so it can be entirely redrawn.

Also, the actions to apply the clip on the cairo_t and get a
cairo_region_t have been split, MechWindow additionally uses this
to set the push_update() region.