Add MechContainer
This is now the parent instance of MechWindow, a MechContainer
represents a toplevel container of areas (eg. with a MechStage),
and is meant to put together processing of updates and event
handling for windows, subsurfaces, and wherever a sub-stage
could make sense.

So far, only the root area creation is delegated on it.