surface: Add surface-type and renderer-type properties
The renderer type is filled in by surface implementations, and
tells the rendering platform being used to draw on the surface.

The surface type is a more generic enum contemplating cases like
"none" and "offscreen", which will be handy as a hint on surface
creation, as eg. "offscreen" could resolve to the best rendering
platform for a given parent.

On the wayland backend, now both SHM and EGL surfaces set the
renderer type accordingly.