surface: Change how surfaces are created altogether
The surface constructor now takes a parent and a surface type,
a surface matching the requirements for that surface type will
be created. MechSurface has been made GInitable altogether, so
it is the caller's responsibility to finish initialization and
fallback on failure.

MechWindow has also gained a renderer-type property, so the
toplevel surface will be updated to match that renderer. As a
consequence, the responsibility to create the toplevel surface
has been moved from MechWindowWayland to MechWindow.

On the wayland side, the ownership of wl_surface structs has moved
from MechWindowWayland to MechSurfaceWayland, this will be useful
when subsurfaces are supported, as wl_subsurfaces are wl_surface
resource. Internally, a wl_surface will only be created for
non-offscreen surfaces.