Add MechGLView
This MechArea requests a GL surface, and internally ensures that every
child area has a MechGLContainer so each child is rendered onto a texture.

This simple implementation does not perform any rendering on the area
surface though, and only does limited input forwarding, which delegates
child picking on child implementations.