Add MechGLBox
This is a MechGLView implementation that performs basic GL rendering
of child textures as quads and implements child picking. By default
children are placed centered on the parent surface, with z=0 and 1:1
x/y dimensions in pixels, the position_child() vmethod can be used
to modify the GL matrix so a children is placed differently.