gl-view: Improve event forwarding
Now keyboard, pointer, and touch IDs get tracked independently.
Additionally to ::pick-child, ::child-coords is now used for the situation
where a child has a grab, and the coordinates move outside that area quad,
so that vmethod is responsible for guessing the right child coordinates in
that situation.

Also, crossing events are generated to ensure the right state of child areas
as the state of a pointer grab changes.