gl-box: Correct handling of inverted textures
When a cairo surface is created on top of a texture, cairo still
seems to assume that the Y-axis ordering is TTB, and not BTT as
GL likes. So instead of inverting the texture on MechGLBox
rendering, set a device transformation so the surface is inverted.

This way, both GL and cairo agree on the way bytes are written,
and mixed rendering can be achieved.