demo: Simplify EGL demo
And rename to gl.c, as there's nothing EGL specific anymore there. Even
though using different EGLContexts is possible, it's not something to
be encouraged really, so take it out of the demo.

The gears area now inherits from MechGLView, so all the GL setup is
done automatically. To add up a bit more than rendering, input events
are handled to rotate the gears.